refactor: extract vminstancenic.go for NIC construction
Move NIC construction functions from vminstance.go to vminstancenic.go:
- createNICOptions struct
- newNetworkInterfaceForVM
- applyTemplateToNic
- createNetworkInterface

Add buildAndCreateNIC orchestrator that consolidates the NIC creation
logic (backend pools, NSG resolution, NIC construction) previously
scattered inline in beginLaunchInstance.
